+++ terraton_fan_app/lib/shared/app_routes.dart
@@ -1,10 +1,11 @@
 // lib/shared/app_routes.dart
 abstract final class AppRoutes {
-  static const splash   = '/splash';
-  static const home     = '/';
-  static const scanQr   = '/scan/qr';
-  static const scanBle  = '/scan/ble';
-  static const nameFan  = '/name-fan';
-  static const control  = '/control';
-  static const settings = '/settings';
+  static const splash             = '/splash';
+  static const home               = '/';
+  static const permissionRequired = '/permission-required';
+  static const scanQr             = '/scan/qr';
+  static const scanBle            = '/scan/ble';
+  static const nameFan            = '/name-fan';
+  static const control            = '/control';
+  static const settings           = '/settings';
 }